doc/src/requirements.txt
funcsigs
mock
PyYaml
sphinx_bootstrap_theme
